requirements.txt
Django==4.1
djangorestframework==3.14.0
pandas==1.5.2
gunicorn
